updating example and refactoring
